requirements.txt
selenium==3.141.0
pytest==6.2.4
pytest-xdist==2.3.0
allure-pytest==2.8.40
allure-python-commons==2.8.40
requests==2.26.0
slackclient==2.9.3
python-dotenv==0.19.0
